requirements.txt
crewai
google-generativeai
qdrant-client
qdrant-client[fastembed]
PyMuPDF
langchain
python-dotenv
pydantic
fastembed
fitz

